move atop tests to test_atop
